refactor(abrg): move the Kyoto result handling to the matching package
Neither function transforms text. One reconciles the results of a two-stage
search and the other re-serializes a structured address, and both are used
only from the two-stage search.
